+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 FaasJS use [Semantic Versioning](https://semver.org/).
 
+[Unreleased]
+
+- `@faasjs/utils`
+  - [Feature] Export more zod types with Zod prefix.
+
 [`v8.0.0-beta.34 (2026-05-26)`](https://github.com/faasjs/faasjs/compare/v8.0.0-beta.33...v8.0.0-beta.34)
 
 - `@faasjs/utils`